fixed the styling of the hr
