+++ src/js/components/aboutTheDataUpdated/AboutTheDataByPage.jsx
@@ -10,7 +10,7 @@ const propTypes = {
     section: PropTypes.string
 };
 
-const AboutTheDataByPage = ({ section }) => <div className="atd__heading">{section.label}</div>;
+const AboutTheDataByPage = ({ section }) => <><div className="atd__heading">{section.label}</div><hr /></>;
 
 AboutTheDataByPage.propTypes = propTypes;
 export default AboutTheDataByPage;
